Рефакторин 01
