+++ composer.json
@@ -1,5 +1,30 @@
 {
-    "require-dev": {
-        "phpunit/phpunit": "6.5.14"
+  "name": "ksnk/templater",
+  "description": "Jinja2 php templater engine",
+  "keywords": ["ksnk", "jinja", "templates"],
+  "license": "MIT",
+  "type": "project",
+  "require": {
+  },
+  "require-dev": {
+    "phpunit/phpunit": "6.5.14"
+  },
+  "autoload": {
+    "psr-4": {
+      "Ksnk\\templater\\": "src/"
     }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "Tests\\": "tests/"
+    }
+  },
+  "config": {
+    "preferred-install": "dist",
+    "sort-packages": true,
+    "optimize-autoloader": true
+  },
+  "minimum-stability": "dev",
+  "prefer-stable": true
+
 }
